backend/requirements.txt
fastapi>=0.100.0,<0.112.0
uvicorn[standard]>=0.20.0,<0.30.0
python-dotenv>=0.21.0,<1.2.0
httpx>=0.25.0,<0.28.0
sqlalchemy>=2.0.0,<2.1.0
asyncpg>=0.27.0,<0.30.0
psycopg2-binary>=2.9.0,<2.10.0
alembic>=1.10.0,<1.14.0
pgvector>=0.4.1,<0.5.0       # For pgvector support with SQLAlchemy
# We'll add more dependencies later
